Factor out pybind11/detail/abi_platform_id.h, type_map.h, cross_extension_shared_state.h from internals.h
IWYU cleanup in the new files and internals.h
