+++ include/pybind11/cast.h
@@ -12,6 +12,7 @@
 
 #include "detail/common.h"
 #include "detail/descr.h"
+#include "detail/native_enum_data.h"
 #include "detail/type_caster_base.h"
 #include "detail/typeid.h"
 #include "pytypes.h"
@@ -58,7 +59,7 @@ public:
 
     template <typename SrcType>
     static handle cast(SrcType &&src, return_value_policy, handle parent) {
-        auto const &natives = native_enum_type_map::get();
+        auto const &natives = cross_extension_shared_states::native_enum_type_map::get();
         auto found = natives.find(std::type_index(typeid(EnumType)));
         if (found != natives.end()) {
             return handle(found->second)(static_cast<Underlying>(src)).release();
@@ -71,7 +72,7 @@ public:
     }
 
     bool load(handle src, bool convert) {
-        auto const &natives = native_enum_type_map::get();
+        auto const &natives = cross_extension_shared_states::native_enum_type_map::get();
         auto found = natives.find(std::type_index(typeid(EnumType)));
         if (found != natives.end()) {
             if (!isinstance(src, found->second)) {
@@ -125,7 +126,7 @@ class type_caster<EnumType,
 
 template <typename T, detail::enable_if_t<std::is_enum<T>::value, int> = 0>
 bool isinstance_native_enum_impl(handle obj, const std::type_info &tp) {
-    auto const &natives = native_enum_type_map::get();
+    auto const &natives = cross_extension_shared_states::native_enum_type_map::get();
     auto found = natives.find(tp);
     if (found == natives.end()) {
         return false;
@@ -1138,7 +1139,9 @@ T cast(const handle &handle) {
                   "Unable to cast type to reference: value is local to type caster");
 #ifndef NDEBUG
     if (is_enum_cast) {
-        if (native_enum_type_map::get().count(std::type_index(typeid(intrinsic_t<T>))) != 0) {
+        if (cross_extension_shared_states::native_enum_type_map::get().count(
+                std::type_index(typeid(intrinsic_t<T>)))
+            != 0) {
             pybind11_fail("Unable to cast native enum type to reference");
         }
     }
