+++ include/pybind11/detail/native_enum_data.h
@@ -5,8 +5,12 @@
 #pragma once
 
 #include "../pytypes.h"
+#include "abi_platform_id.h"
 #include "common.h"
+#include "cross_extension_shared_state.h"
+#include "type_map.h"
 
+#include <string>
 #include <typeindex>
 
 PYBIND11_NAMESPACE_BEGIN(PYBIND11_NAMESPACE)
@@ -56,4 +60,28 @@ public:
 };
 
 PYBIND11_NAMESPACE_END(detail)
+
+PYBIND11_NAMESPACE_BEGIN(cross_extension_shared_states)
+
+struct native_enum_type_map_v1_adapter {
+    static constexpr const char *abi_id() {
+        return "__pybind11_native_enum_type_map_v1" PYBIND11_PLATFORM_ABI_ID_V4 "__";
+    }
+
+    using payload_type = detail::type_map<PyObject *>;
+
+    static void payload_clear(payload_type &payload) {
+        for (auto it : payload) {
+            Py_DECREF(it.second);
+        }
+        payload.clear();
+    }
+};
+
+using native_enum_type_map_v1
+    = detail::cross_extension_shared_state<native_enum_type_map_v1_adapter>;
+using native_enum_type_map = native_enum_type_map_v1;
+
+PYBIND11_NAMESPACE_END(cross_extension_shared_states)
+
 PYBIND11_NAMESPACE_END(PYBIND11_NAMESPACE)
