+++ tests/extra_python_package/test_files.py
@@ -47,13 +47,16 @@ main_headers = {
 }
 
 detail_headers = {
+    "include/pybind11/detail/abi_platform_id.h",
     "include/pybind11/detail/class.h",
     "include/pybind11/detail/common.h",
+    "include/pybind11/detail/cross_extension_shared_state.h",
     "include/pybind11/detail/descr.h",
     "include/pybind11/detail/init.h",
     "include/pybind11/detail/internals.h",
     "include/pybind11/detail/native_enum_data.h",
     "include/pybind11/detail/type_caster_base.h",
+    "include/pybind11/detail/type_map.h",
     "include/pybind11/detail/typeid.h",
 }
